+++ MXP3/include/scene.h
@@ -2,13 +2,14 @@
 
 #include "entity.h"
 #include "renderer.h"
+#include "input.h"
 
 class __declspec(dllexport) Scene : public Entity {
 public:
 	Scene(SDL_Renderer* renderer);
 	virtual ~Scene();
 
-	void update(float deltatime);
+	void update(double deltatime);
 
 	Entity* pic;
